Rename common_utils folder to utils
