+++ framework/test/jest/unit/specs/modules/chain/dpos/verifyBlockForger.spec.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const { Dpos, Slots } = require('../../../../../../../src/modules/chain/dpos');
-const { constants } = require('../../../../../common_utils');
+const { constants } = require('../../../../../utils');
 const { delegatePublicKeys } = require('./round_delegates');
 
 describe('dpos.verifyBlockForger()', () => {
